chore: move header around a bit
