+++ include/loader/MoreStringFunctions.h
